requirements.txt
pyyaml==5.3.1
docker==4.2.1
requests
PyGithub
Jinja2
gitpython==3.1.3
python-dotenv

-r controller/requirements.txt